chore(ui): replace old promotion modal and add abort promotion in new one
